move constants.ts into src/runtime
