+++ src/runtime/internal/transitions.ts
@@ -5,7 +5,7 @@ import { create_rule, delete_rule } from './style_manager';
 import { custom_event } from './dom';
 import { add_render_callback } from './scheduler';
 import { TransitionConfig } from '../transition';
-import { resolved_promise } from '../../constants';
+import { resolved_promise } from '../constants';
 
 let promise: Promise<void>|null;
