+++ src/runtime/motion/spring.ts
@@ -1,7 +1,7 @@
 import { Readable, writable } from 'svelte/store';
 import { loop, now, Task } from 'svelte/internal';
 import { is_date } from './utils';
-import { resolved_promise } from '../../constants';
+import { resolved_promise } from '../constants';
 
 interface TickContext<T> {
 	inv_mass: number;
